pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-agent"
version = "0.1.0"
description = "CLI tool to generate LinkedIn posts, images, and comments using AI"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "openai>=1.0.0",
    "Jinja2>=3.1.0",
    "rich>=13.0.0",
    "pyperclip>=1.8.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
agent = "linkedin_agent.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
